Updated api to include a v1 and shared structure to maintain first api for beta-testers
