+++ pages/api/extension/v1/new-tab.js
@@ -1,0 +1,4 @@
+// V1 endpoint for production extension - uses shared handler
+import handler from "../_shared/new-tab-handler";
+
+export default handler;
